Adds step that executes to end but with bugs
Bug#1: The order of the steps is wrong. Should move all the
files to the temp directory before updating the project xml.

Bug#2: The update to the project xml file duplicates the root
node and all it's children. Essentially duplicating the file.

Will fix bug#1 by reordering the operations, copying files to
temp directory first and then updating the project xml file
in the temp directory.